Added two build flavors: local, and remote. When debugging locally, use local.
I did this so we could simplify switching between the remote and local url.
If you want to debug locally, enter your ip address in the file
app/src/local/res/values/strings.xml
The app uses those strings for addresses when it is in the local build flavor.

If you want to run the app to connect to the remote server, simply switch to the
remote build flavor.